projeto 99% convertido em  class
