--- Semana19/Projeto/src/business/unFriendBusiness/unFriendBusiness.ts
@@ -1,9 +1,0 @@
-import deleteFriend from "../../data/deleteFriend/deleteFriend";
-
-export const unFriendBusiness = async (id: any) => {
-    try {
-        await deleteFriend(id)
-    } catch (error) {
-        throw new Error(error);
-    }
-}
