--- Semana19/Projeto/src/data/insertUser/insertUser.ts
@@ -1,20 +1,0 @@
-import { connection } from "../../connection/connection";
-
-export default async function insertUser(
-    id: string,
-    name: string,
-    email: string,
-    password: string
-) {
-    try {
-        await connection('labook_users')
-            .insert({
-                id,
-                name,
-                email,
-                password
-            })
-    } catch (error) {
-        throw new Error(error.sqlMessage || error.message);
-    }
-}
