--- Semana19/Projeto/src/data/loginUser/loginUser.ts
@@ -1,15 +1,0 @@
-import { connection } from "../../connection/connection";
-
-export default async function loginUser(
-    email: string
-): Promise<any> {
-    try {
-        const queryResult: any = await connection("labook_users")
-            .select("*")
-            .where({ email })
-
-        return queryResult[0]
-    } catch (error) {
-        throw new Error(error.sqlMessage || error.message);
-    }
-}
