--- Semana19/Projeto/src/data/selectPost/selectPost.ts
@@ -1,16 +1,0 @@
-import { connection } from "../../connection/connection";
-
-export default async function selectPost(
-    id: string
-) {
-    try {
-        const queryResult: any = await connection("labook_posts")
-            .select("*")
-            .where({ id })
-            
-        return queryResult[0]
-
-    } catch (error) {
-        throw new Error(error.sqlMessage || error.message);
-    }
-}
