+++ Semana19/Projeto/src/index.ts
@@ -3,7 +3,7 @@
 import express, { Express } from "express"
 import cors from "cors"
 import { userRoutes } from "./routes/userRoutes"
-import { inviteFriend } from "./controller/inviteFriend/inviteFriend"
+import { postRoutes } from "./routes/postRoutes"
 
 /**************************** CONFIG ******************************/
 
@@ -11,8 +11,8 @@ const app: Express = express()
 app.use(express.json())
 app.use(cors())
 app.use("/user", userRoutes)
-app.use("/post", userRoutes)
 app.use("/friends", userRoutes)
+app.use("/post", postRoutes)
 
 /**************************** SERVER INIT ******************************/
