+++ Semana19/Projeto/src/routes/friendsRouter.ts
@@ -1,0 +1,8 @@
+
+import express from 'express'
+import { friendsController } from '../controller/friendsController/FriendsController'
+
+export const userRoutes = express.Router()
+
+userRoutes.post("/:id", friendsController.inviteFriend)
+userRoutes.delete("/:id", friendsController.unFriend)
